+++ assets/javascripts/kitten/components/information/icon-badge/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import classNames from 'classnames'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import deprecated from 'prop-types-extra/lib/deprecated'
 import { pxToRem } from '../../../helpers/utils/typography'
 
 const StyledBadge = styled.span`
@@ -94,11 +93,7 @@ const StyledBadge = styled.span`
 export const IconBadge = ({
   className,
   children,
-  disabled,
-  valid,
   empty,
-  big,
-  huge,
   size,
   border,
   backgroundColor,
@@ -113,11 +108,7 @@ export const IconBadge = ({
       `k-IconBadge--${size}`,
       `k-IconBadge--${status}`,
       {
-        'k-IconBadge--disabled': disabled && status === 'info', // if default prop
-        'k-IconBadge--valid': valid && status === 'info', // if default prop
         'k-IconBadge--empty': empty,
-        'k-IconBadge--big': big,
-        'k-IconBadge--huge': huge,
         'k-IconBadge--hasBorderStyles': border.length > 0,
         'k-IconBadge--hasBorder': hasBorder,
       },
@@ -145,11 +136,7 @@ IconBadge.defaultProps = {
 }
 
 IconBadge.propTypes = {
-  disabled: deprecated(PropTypes.bool, 'Use status="disabled" instead'),
-  valid: deprecated(PropTypes.bool, 'Use status="success" instead'),
   empty: PropTypes.bool,
-  big: deprecated(PropTypes.bool, 'Use `size` prop instead.'),
-  huge: deprecated(PropTypes.bool, 'Use `size` prop instead.'),
   size: PropTypes.oneOf(['micro', 'tiny', 'normal', 'big', 'huge']),
   backgroundColor: PropTypes.string,
   border: PropTypes.shape({