+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `IconBadge`: Remove `disabled`, `valid`, `big` and `huge` deprecated props.
 - Feature: `TextInput`: Add `autoresize` value to `tag` prop.
 - Fix: `*EntryTable`: Fix column value for attribute `scope`.